+++ src/main/java/by/bsu/wialontransport/kafka/transportable/data/TransportableSavedData.java
@@ -20,7 +20,7 @@ public final class TransportableSavedData extends TransportableData {
                                   final double latitude,
                                   final double longitude,
                                   final int course,
-                                  final int speed,
+                                  final double speed,
                                   final int altitude,
                                   final int amountOfSatellites,
                                   final double reductionPrecision,
